+++ auto/src/not_mint.c
@@ -1,7 +1,9 @@
 #include <mint.h>
+#include <stddef.h>
 
 mint *not_mint (mint *numin){
-  unsigned long size = size_mint(numin);
+  // unsigned long size = size_mint(numin);
+  size_t size = size_mint(numin);
   mint *numout = make_mint(size);
   not_mint_manually(numin, numout);
   return numout;
