+++ manual/src/or_mint_manually.c
@@ -2,11 +2,10 @@
 #define max(a, b) (a)<(b)?(b):(a)
 
 void or_mint_manually (mint *numa, mint *numb, mint *numout){
-  // unsigned long sizea = size_mint(numa);
-  // unsigned long sizeb = size_mint(numb);
-  // unsigned long size = max(sizea, sizeb);
-  unsigned long size = real_size_mint(numout);
-  unsigned long index;
+  // unsigned long size = real_size_mint(numout);
+  // unsigned long index;
+  size_t size = real_size_mint(numout);
+  size_t index;
   for (index = 0; index < size; index++){
     mint_cell numca = get_mint(index, 0, numa);
     mint_cell numcb = get_mint(index, 0, numb);
