+++ manual/src/shift_size_mint.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <mint.h>
+#include <stddef.h>
 
-extern unsigned long seek_size_mint (mint*);
-extern unsigned long offset_size_mint (mint*);
+extern size_t seek_size_mint (mint*);
+extern size_t offset_size_mint (mint*);
